Refactor replicated job orchestrator and add initialization
Refactors the replicated job orchestrator to make testing simpler, and
then adds initialization logic to it.
