+++ manager/orchestrator/replicatedjob/replicatedjob_suite_test.go
@@ -1,4 +1,4 @@
-package job_test
+package replicatedjob
 
 import (
 	"testing"
